refactor tests into separate modules
